+++ apps/pwa/CHANGELOG.md
@@ -249,7 +249,7 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ### Bug Fixes
 
-* remove nestjs-authentication @generated/all dependency from package.json ([#168](https://github.com/tractr/stack/issues/168)) ([20891bc](https://github.com/tractr/stack/commit/20891bce7697ef834835dff29f24313ac3a09b51))
+* remove nestjs-authentication @tractr/generated-all dependency from package.json ([#168](https://github.com/tractr/stack/issues/168)) ([20891bc](https://github.com/tractr/stack/commit/20891bce7697ef834835dff29f24313ac3a09b51))
 
 ## [1.6.1](https://github.com/tractr/stack/compare/pwa-1.6.0...pwa-1.6.1) (2021-07-07)